Added more UnitTest module for Rayguns update file structure to match with Michale
Added testing for:
-FireAt_AlienTest - loaded gun, dodging alien
-FireAt_AlienTest - empty gun, dodging alien
-FireAt_AlienTest - loaded gun, incapacitated alien (no dodge)
-FireAt_AlienTest - loaded gun, incapacitated alien with 1 HP, expect to see death
