+++ CSD_Aliens.iml
@@ -5,8 +5,8 @@
     <content url="file://$MODULE_DIR$">
       <sourceFolder url="file://$MODULE_DIR$/src" isTestSource="false" />
       <sourceFolder url="file://$MODULE_DIR$/RaygunTests" isTestSource="true" />
-      <sourceFolder url="file://$MODULE_DIR$/Raygun Test" isTestSource="true" />
-      <sourceFolder url="file://$MODULE_DIR$/Raygun Test/resources" type="java-test-resource" />
+      <sourceFolder url="file://$MODULE_DIR$/test" isTestSource="true" />
+      <sourceFolder url="file://$MODULE_DIR$/test/resources" type="java-test-resource" />
     </content>
     <orderEntry type="inheritedJdk" />
     <orderEntry type="sourceFolder" forTests="false" />
